+++ android/src/main/kotlin/com/gdelataillade/alarm/services/AudioService.kt
@@ -1,8 +1,11 @@
 package com.gdelataillade.alarm.services
 
 import android.content.Context
+import android.media.AudioAttributes
+import android.media.AudioManager
 import android.media.MediaPlayer
 import android.media.RingtoneManager
+import android.os.Build
 import com.gdelataillade.alarm.models.VolumeFadeStep
 import java.util.concurrent.ConcurrentHashMap
 import java.util.Timer